Fixing lint and getting rid of openai
