+++ env.sample
@@ -1,5 +1,6 @@
-VITE_OPENAI_API_KEY=YOUR_API_KEY
 VITE_API_URL="https://api.wsuv-hp-capstone.com"
 VITE_API_PORT=443
-// localhost
-// 5040
+
+// For development only:
+// VITE_API_URL=localhost
+// VITE_API_PORT=5040
