+++ package.json
@@ -14,7 +14,6 @@
     "chartjs-plugin-zoom": "^2.2.0",
     "core-js": "^3.37.1",
     "mkdirp": "^3.0.1",
-    "openai": "^4.91.1",
     "roboto-fontface": "*",
     "vue": "^3.4.31",
     "vue-chart-3": "^3.1.8",
