+++ src/components/PrintJobs.vue
@@ -29,7 +29,7 @@ import { useRouter } from 'vue-router';
 import ModuleToolbar from "./module-toolbar.vue";
 import PrintJobCreate from "./PrintJob/print-job-create.vue";
 import ErrorPopups from "./ErrorPopups.vue";
-import ChatWidget from './ChatBot/ChatWidget.vue';
+import ChatWidget from './ChatBot/chat-widget.vue';
 
 const errorMessage = ref('');
 const router = useRouter();
