+++ src/components/WorkFlows.vue
@@ -29,7 +29,7 @@ import { getCollection, formatSteps, postWorkflow } from "./api.js";
 import ModuleToolbar from "./module-toolbar.vue";
 import WorkflowCreate from "./Workflow/workflow-create.vue";
 import ErrorPopups from "./ErrorPopups.vue";
-import ChatWidget from './ChatBot/ChatWidget.vue';
+import ChatWidget from './ChatBot/chat-widget.vue';
 
 const errorMessage = ref('');
 const router = useRouter();
